+++ examples/boolean.rs
@@ -14,4 +14,8 @@ fn main() {
     envmnt::set("MY_ENV_VAR", "SOME VALUE");
     let same = envmnt::is_equal("MY_ENV_VAR", "SOME VALUE");
     println!("Value Is Same: {}", &same);
+    let mut contains = envmnt::contains("MY_ENV_VAR", "_ENV_");
+    println!("Value Contained: {}", &contains);
+    contains = envmnt::contains_ignore_case("MY_ENV_VAR", "_env_");
+    println!("Value Contained (case insensitive): {}", &contains);
 }